Make actually used fovutls functions methods of FOV
